factored out friendform component
